refactor(proxy): consolidate proxy package files (22 → 10)
Mechanical mergers only. No logic changes, no public API changes, no
test changes.

| New file        | Sources                                                                                                                              |
|-----------------|--------------------------------------------------------------------------------------------------------------------------------------|
| manager.go      | manager.go + manager_writepath.go + dashboard.go + errors.go                                                                         |
| multipart.go    | multipart.go + multipart_lock.go + multipart_encrypt.go + multipart_parts.go + multipart_cleanup.go + multipart_complete.go          |
| cache.go        | cache.go + cache_tee.go                                                                                                              |
| objects_read.go | objects.go + objects_read.go + integrity.go                                                                                          |
| objects_write.go| objects_write.go + copy_materialize.go                                                                                               |

Unchanged: core.go, coordinator.go, reconcile.go, lifecycle.go,
encryption_helpers.go.

22 → 10 production .go files (−12, −55%).

Closes #842
