+++ .version
@@ -1,1 +1,1 @@
-v0.47.8
+v0.47.9
